+++ app/src/main/java/ru/whiteleaf/notes/presentation/note_list/NoteListViewModel.kt
@@ -1,6 +1,6 @@
 package ru.whiteleaf.notes.presentation.note_list
 
-import androidx.fragment.app.FragmentActivity
+import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -16,13 +16,17 @@ import ru.whiteleaf.notes.domain.use_case.RenameNoteUseCase
 import ru.whiteleaf.notes.domain.use_case.RenameNotebookUseCase
 import kotlinx.coroutines.launch
 import ru.whiteleaf.notes.domain.interactor.PreferencesInteractor
-import ru.whiteleaf.notes.domain.repository.SecurityPreferences
-import ru.whiteleaf.notes.domain.use_case.CheckNotebookAccessUseCase
-import ru.whiteleaf.notes.domain.use_case.ClearNotebookKeysUseCase
+import ru.whiteleaf.notes.domain.repository.AuthenticationRequiredException
+import ru.whiteleaf.notes.domain.use_case.CreateKeyForNotebookUseCase
+import ru.whiteleaf.notes.domain.use_case.DecryptNotebookUseCase
+import ru.whiteleaf.notes.domain.use_case.DeleteKeyForNotebookUseCase
 import ru.whiteleaf.notes.domain.use_case.EncryptNotebookUseCase
-import ru.whiteleaf.notes.domain.use_case.LockNotebookUseCase
+import ru.whiteleaf.notes.domain.use_case.IsNotebookProtectedUseCase
+import ru.whiteleaf.notes.domain.use_case.IsNotebookUnlockedUseCase
 import ru.whiteleaf.notes.domain.use_case.UnlockNotebookUseCase
+import ru.whiteleaf.notes.domain.use_case.LockNotebookUseCase
 import java.io.IOException
+import java.security.InvalidKeyException
 
 class NoteListViewModel(
     private val getNotesUseCase: GetNotesUseCase,
@@ -33,12 +37,14 @@ class NoteListViewModel(
     private val renameNotebookUseCase: RenameNotebookUseCase,
     private val shareNotebookUseCase: ShareNotebookUseCase,
     private val deleteNotebookUseCase: DeleteNotebookByPathUseCase,
-    private val encryptNotebookUseCase: EncryptNotebookUseCase,
-    private val clearNotebookKeys: ClearNotebookKeysUseCase,
+    private val isNotebookProtectedUseCase: IsNotebookProtectedUseCase,
+    private val isNotebookUnlockedUseCase: IsNotebookUnlockedUseCase,
     private val unlockNotebookUseCase: UnlockNotebookUseCase,
-    private val checkNotebookAccessUseCase: CheckNotebookAccessUseCase,
     private val lockNotebookUseCase: LockNotebookUseCase,
-    private val securityPreferences: SecurityPreferences,
+    private val createKeyForNotebookUseCase: CreateKeyForNotebookUseCase,
+    private val deleteKeyForNotebookUseCase: DeleteKeyForNotebookUseCase,
+    private val encryptNotebookUseCase: EncryptNotebookUseCase,
+    private val decryptNotebookUseCase: DecryptNotebookUseCase,
     private val preferencesInteractor: PreferencesInteractor,
     private val notebookPath: String?
 ) : ViewModel() {
@@ -52,26 +58,12 @@ class NoteListViewModel(
     private val _navigationEvent = MutableLiveData<NavigationEvent>()
     val navigationEvent: LiveData<NavigationEvent> = _navigationEvent
 
-    // Новые LiveData для безопасности
-    private val _notebookSecurityState = MutableLiveData<NotebookSecurityState>()
-
-    private val _authenticationRequired = MutableLiveData<Boolean>()
-    val authenticationRequired: LiveData<Boolean> = _authenticationRequired
-
-    private val _encryptionResult = MutableLiveData<Result<Unit>>()
-    val encryptionResult: LiveData<Result<Unit>> = _encryptionResult
-
-    private var isEncrypted =
-        notebookPath?.let { checkNotebookAccessUseCase.isNotebookEncrypted(it) } == true
-    private var hasAccess = true
-
-    private val _isPlannerView = MutableLiveData<Boolean>(false)
+    private val _isPlannerView = MutableLiveData(false)
 
     init {
         loadViewMode()
         loadNotes()
         saveLastOpenedNotebook()
-        checkSecurityState()
     }
 
     private fun loadViewMode() {
@@ -87,40 +79,29 @@ class NoteListViewModel(
         _navigationEvent.postValue(NavigationEvent.NavigateToNotebook(notebookPath))
     }
 
-    fun getViewMode():Boolean {
-        return _isPlannerView.value?:false
-    }
-
-    private fun checkSecurityState() {
-
-        viewModelScope.launch {
-            isEncrypted =
-                notebookPath?.let { checkNotebookAccessUseCase.isNotebookEncrypted(it) } == true
-            hasAccess = notebookPath?.let { checkNotebookAccessUseCase(it) } != false
-
-            _notebookSecurityState.postValue(///
-                NotebookSecurityState(
-                    isEncrypted = isEncrypted,
-                    isUnlocked = hasAccess,
-                    requiresAuthentication = isEncrypted && !hasAccess
-                )
-            )
-            _authenticationRequired.postValue(isEncrypted && !hasAccess)///
-        }
+    fun getViewMode(): Boolean {
+        return _isPlannerView.value ?: false
     }
 
     fun loadNotes() {
-        _noteListState.postValue(NoteListState.Loading)
-
         viewModelScope.launch {
-            if (notebookPath != null) {
-                val hasAccess = checkNotebookAccessUseCase(notebookPath)
-                if (!hasAccess) {
-                    _noteListState.postValue(NoteListState.Blocked)
-                    return@launch
-                }
-            }
+            _noteListState.postValue(NoteListState.Loading)
+            var isProtected = false
+
             try {
+                if (notebookPath != null) {
+
+                    isProtected = isNotebookProtectedUseCase(notebookPath)
+                    val isUnlocked = isNotebookUnlockedUseCase(notebookPath)
+
+                    if (isProtected && !isUnlocked) {
+                        _noteListState.postValue(NoteListState.Blocked)
+                        _navigationEvent.postValue(NavigationEvent.ShowBiometric)
+                        return@launch
+                    }
+                }
+
+                println("DEBUG: NoteListViewmodel: Загрузка заметок, isProtected= $isProtected")
                 val notesList = getNotesUseCase(notebookPath)
 
                 notesList.forEach { note ->
@@ -130,71 +111,101 @@ class NoteListViewModel(
                     }
                 }
                 _noteListState.postValue(
-                    NoteListState.Success(
-                        isEncrypted,
-                        notesList.filter { it.isNotEmpty() })
+                    NoteListState.Success(isProtected, notesList.filter { it.isNotEmpty() })
                 )
 
+            } catch (e: AuthenticationRequiredException) {
+                _noteListState.postValue(NoteListState.Blocked)
             } catch (e: IOException) {
                 _noteListState.postValue(NoteListState.Error("Ошибка загрузки заметок: ${e.message}"))
             } catch (e: Exception) {
-                showMessage("Неизвестная ошибка: ${e.message}")
-                _noteListState.postValue(NoteListState.Error("Неизвестная ошибка: ${e.message}"))
+                if (e.cause is InvalidKeyException) {
+                    _noteListState.postValue(NoteListState.Blocked)
+                    if (notebookPath != null) lockNotebookUseCase(notebookPath)
+                    println("DEBUG: NoteListViewmodel: InvalidKeyException ${e.message}")
+                } else
+                    _noteListState.postValue(NoteListState.Error(e.message ?: "Ошибка загрузки"))
             } finally {
-                println("Окончание загрузки заметок")
+                println("DEBUG: NoteListViewmodel: Окончание загрузки заметок")
             }
         }
     }
 
-    fun encryptNotebook() {
-        viewModelScope.launch {
-            if (notebookPath != null) {
-                // Проверяем, не зашифрован ли уже блокнот
-                if (securityPreferences.isNotebookEncrypted(notebookPath)) {
-                    showMessage("Блокнот уже зашифрован")
-                    return@launch
-                }
+    fun unlockNotebook(context: Context) {
+        val isProtected =
+            if (notebookPath != null) isNotebookProtectedUseCase(notebookPath) else false //encryptionRepository.hasKey(notebookPath) else false
+        println("DEBUG: unlock notebook, notebookPath = $notebookPath, isProtected = $isProtected")
 
-                encryptNotebookUseCase(notebookPath)
-                    .onSuccess {
-                        showMessage("Блокнот успешно зашифрован")
-                        checkSecurityState()
-                        loadNotes() // Перезагружаем чтобы показать заблокированное состояние
-                    }
-                    .onFailure { error ->
-                        val errorMessage = error.message ?: "Неизвестная ошибка"
-                        showMessage("Ошибка шифрования: $errorMessage")
-                        println("❌ Ошибка в ViewModel: $errorMessage")
-                        error.printStackTrace()
-                    }
-            }
-        }
-    }
-
-    fun unlockNotebook(activity: FragmentActivity) {
-        viewModelScope.launch {
-            if (notebookPath != null) {
-
-                unlockNotebookUseCase(notebookPath, activity).onSuccess {
-                    showMessage("Записная книжка разблокирована")
-                    checkSecurityState()
-                    loadNotes() // Загружаем расшифрованные заметки
-                }.onFailure { error ->
-                    showMessage(error.message.toString())
-                    println("❌ Ошибка разблокировки: ${error.message}")
-                    error.printStackTrace()
+        if (notebookPath != null && isProtected)
+            viewModelScope.launch {
+                val unlocked = unlockNotebookUseCase(notebookPath, context)
+                if (unlocked) {
+                    println("DEBUG: unlock notebook: success")
+                    loadNotes()
+                } else {
+                    _noteListState.postValue(NoteListState.Blocked)
                 }
             }
-        }
     }
 
     fun lockNotebook() {
-        showMessage("Шифруем")
         if (notebookPath != null) {
             lockNotebookUseCase(notebookPath)
-            showMessage("Записная книжка зашифрована")
-            checkSecurityState()
-            loadNotes() // Обновляем список заметок (должен стать пустым)
+            _noteListState.postValue(NoteListState.Blocked)
+        }
+    }
+
+    fun encryptNotebook(context: Context) {
+        if (notebookPath != null) viewModelScope.launch {
+            try {
+                if (isNotebookProtectedUseCase(notebookPath)) {
+                    _noteListState.value = NoteListState.Error("Записная книжка уже защищищена")
+                    return@launch
+                }
+                createKeyForNotebookUseCase(notebookPath)
+                unlockNotebookUseCase(
+                    notebookPath,
+                    context,
+                    title = "Защита записной кникки",
+                    reason = "Для защиты"
+                )
+                encryptNotebookUseCase(notebookPath)
+                loadNotes()
+                _message.value = "Записная книжка зашифрована"
+            } catch (e: AuthenticationRequiredException) {
+                _noteListState.value =
+                    NoteListState.Error("Не удалось зашифровать записную книжку: ${e.message}")
+                //_noteListState.value = NoteListState.Blocked
+                println("DEBUG: NoteListVM fail encrypt notebook: ${e.message}")
+
+            } catch (e: Exception) {
+                _noteListState.value = NoteListState.Error("Ошибка шифрования: ${e.message}")
+                println("DEBUG: Ошибка шифрования: ${e.message}")
+            }
+        }
+    }
+
+    fun decryptNotebook(context: Context) {
+        if (notebookPath != null) viewModelScope.launch {
+            try {
+                if (!isNotebookProtectedUseCase(notebookPath)) {
+                    _noteListState.value = NoteListState.Error("Защита не установлена")
+                    return@launch
+                }
+                val unlocked =
+                    unlockNotebookUseCase(notebookPath, context, reason = "Для снятия защиты")
+                if (!unlocked) {
+                    _noteListState.value =
+                        NoteListState.Error("Не удалось подтвердить личность")
+                    return@launch
+                }
+                decryptNotebookUseCase(notebookPath)
+                deleteKeyForNotebookUseCase(notebookPath)
+                loadNotes()
+                _message.value = "Защита записной книжки снята"
+            } catch (e: Exception) {
+                _noteListState.value = NoteListState.Error("Ошибка расшифровки: ${e.message}")
+            }
         }
     }
 
@@ -249,61 +260,79 @@ class NoteListViewModel(
         }
     }
 
-    fun renameNotebook(newName: String) {
-        viewModelScope.launch {
-            try {
-                if (newName != notebookPath && notebookPath != null) {
-                    renameNotebookUseCase(notebookPath, newName)
-                    showMessage("Название записной книжки изменено")
-                    _navigationEvent.postValue(NavigationEvent.NavigateToNotebook(newName))
-                } else showMessage("Ошибка переименования")
-            } catch (e: Exception) {
-                showMessage("Ошибка переименования: ${e.message}")
+    fun renameNotebook(newName: String, context: Context) {
+        if (notebookPath != null)
+            viewModelScope.launch {
+                try {
+                    val unlocked =
+                        unlockNotebookUseCase(notebookPath, context, reason = "Для переименования")
+
+                    if (!unlocked) {
+                        _message.postValue("Не удалось подтвердить личность")
+                        return@launch
+                    }
+
+                    if (newName != notebookPath) {
+                        renameNotebookUseCase(notebookPath, newName)
+                        showMessage("Название записной книжки изменено")
+                        _navigationEvent.postValue(NavigationEvent.NavigateToNotebook(newName))
+                    } else showMessage("Ошибка переименования")
+                } catch (e: AuthenticationRequiredException) {
+                    _message.postValue("Записная книжка заблокирована: ${e.message}")
+                } catch (e: Exception) {
+                    showMessage("Ошибка переименования: ${e.message}")
+                }
             }
-        }
     }
 
-    fun shareNotebook() {
+    fun shareNotebook(context: Context) {
         viewModelScope.launch {
             if (notebookPath != null)
                 try {
-                    val result = shareNotebookUseCase(notebookPath)
+                    val unlocked =
+                        unlockNotebookUseCase(notebookPath, context, reason = "Для экспорта")
+
+                    if (!unlocked) {
+                        _message.postValue("Не удалось подтвердить личность")
+                        return@launch
+                    }
+
+                    val result = shareNotebookUseCase(notebookPath, "123") //todo пароль спрашивать
 
                     if (result.isSuccess)
                         _navigationEvent.postValue(NavigationEvent.ExportLink(result.getOrNull()))
                     else
                         showMessage(result.exceptionOrNull()?.message ?: "Unknown error")
 
+                } catch (e: AuthenticationRequiredException) {
+                    _message.postValue("Записная книжка заблокирована: ${e.message}")
                 } catch (e: Exception) {
                     showMessage("Ошибка передачи файла записной книжки: ${e.message}")
                 }
         }
     }
 
-    fun deleteNotebook() {
-        viewModelScope.launch {
-            try {
-                if (notebookPath != null) {
-//                    println("🗑️ НАЧАЛО УДАЛЕНИЯ БЛОКНОТА: $notebookPath")
-//                    println("🔑 Очищаем ключи...")
-//                    clearNotebookKeys(notebookPath)
-//
-//                    println("📊 Сбрасываем состояние безопасности...")
-//                    securityPreferences.setNotebookEncrypted(notebookPath, false)
-//                    securityPreferences.setNotebookUnlocked(notebookPath, false)
-//
-//                    println("📁 Удаляем файлы блокнота...")
+    fun deleteNotebook(context: Context) {
+        if (notebookPath != null)
+            viewModelScope.launch {
+                try {
+                    val unlocked =
+                        unlockNotebookUseCase(notebookPath, context, reason = "Для удаления")
+
+                    if (!unlocked) {
+                        _message.postValue("Не удалось подтвердить личность")
+                        return@launch
+                    }
 
                     deleteNotebookUseCase(notebookPath)
-                    //println("✅ Блокнот полностью удален")
+
                     _navigationEvent.postValue(NavigationEvent.NavigateUp)
                     showMessage("Записная книжка удалена")
 
-                } else showMessage("Ошибка удаления записной книжки: путь не задан")
-            } catch (e: Exception) {
-                showMessage("Ошибка удаления записной книжки: ${e.message}")
+                } catch (e: Exception) {
+                    showMessage("Ошибка удаления записной книжки: ${e.message}")
+                }
             }
-        }
     }
 
     fun onNoteClicked(noteId: String) =
@@ -318,11 +347,17 @@ class NoteListViewModel(
     }
 
     fun onNotebookExited(toNote: Boolean) {
-        if (toNote) return
-        if (notebookPath != null) {
-            println("🔒 Блокируем блокнот при выходе: $notebookPath")
-            lockNotebookUseCase(notebookPath)
-        }
+        println("DEBUG: Выход из блокнота $notebookPath, toNote=$toNote")
+        val state = _noteListState.value
+
+        if (toNote || state !is NoteListState.Success) return
+        println("DEBUG: isEncrypted = ${state.isEncrypted} ")
+
+        if (notebookPath != null)
+            if (state.isEncrypted) {
+                println("DEBUG: убираем признак разблокировки при выходе: $notebookPath")
+                lockNotebookUseCase(notebookPath)
+            }
     }
 
     override fun onCleared() {
